chore(style): remove old style casts and cross sign comparisons
Enable warnings for both.

Fix #6010
Fix #6012

+++ test/core/toxstring_test.cpp
@@ -41,13 +41,13 @@ private:
     static const QString testStr;
     static const QByteArray testByte;
     static const uint8_t* testUINT8;
-    static const size_t lengthUINT8;
+    static const int lengthUINT8;
 
     //"" - empty test text
     static const QString emptyStr;
     static const QByteArray emptyByte;
     static const uint8_t* emptyUINT8;
-    static const size_t emptyLength;
+    static const int emptyLength;
 };
 
 
@@ -56,13 +56,13 @@ private:
 const QString TestToxString::testStr = QStringLiteral("My Test String");
 const QByteArray TestToxString::testByte = QByteArrayLiteral("My Test String");
 const uint8_t* TestToxString::testUINT8 = reinterpret_cast<const uint8_t*>("My Test String");
-const size_t TestToxString::lengthUINT8 = 14;
+const int TestToxString::lengthUINT8 = 14;
 
 //"" - empty test text
 const QString TestToxString::emptyStr = QStringLiteral("");
 const QByteArray TestToxString::emptyByte = QByteArrayLiteral("");
 const uint8_t* TestToxString::emptyUINT8 = reinterpret_cast<const uint8_t*>("");
-const size_t TestToxString::emptyLength = 0;
+const int TestToxString::emptyLength = 0;
 
 /**
  * @brief Use QString as input data, check output: QString, QByteArray, size_t and uint8_t
